backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]
sqlalchemy[asyncio]==2.0.36
asyncpg
alembic
pydantic==2.10.0
pydantic-settings
PyJWT[crypto]
passlib[bcrypt]
redis
celery
httpx
python-multipart
boto3
pytest
pytest-asyncio
aiosqlite
factory-boy
paho-mqtt
